Change home page title.
Made typing animation go through only 1 string as opposed to 3 and
changed the color of the word "noun" to emphasize it more. This was
changed because the animation was too long to be played each time the
user would return to the home page from another page in the website.