Change death handling to restart game when any ONE snake has died.
Before, the game would continue even if a single Snake died, however,
this caused bugs in the game's grid system and made the game unplayable
after a Snake died a few times. As a result, the design decision was
made to retart the entire game once a single Snake has died to keep the
game playable.